requirements.txt
PyMuPDF
google-generativeai
python-telegram-bot
